feat(rules): create-rule shortcut from transactions table (#93)
Closes #93.

Adds a Sparkles row action on the transactions table: "Create rule from
this transaction". Clicking opens the same RuleFormModal used on /rules,
pre-filled with:
- pattern: merchant_name → description_clean → description (first
  non-empty wins)
- match_type: contains
- category_id: the row's current category (user can change)
- priority: max(existing rules) + 10

After submit success, a bottom-right toast offers two follow-up actions:
- "View rules" → Link to /rules
- "Apply to existing transactions" → POST /categorization-rules/apply
  and re-fetch the transactions list so new categorization shows
  immediately

Refactor: extracts RuleFormModal from pages/RulesPage.tsx into
components/RuleFormModal.tsx and extends its `defaults` prop to carry
pattern, match_type, category_id, priority, and is_active. RulesPage
keeps its existing UX unchanged.
